+++ Sources/MapboxNavigationUIKit/CarPlay/CarPlayMapViewController.swift
@@ -46,8 +46,8 @@ open class CarPlayMapViewController: UIViewController {
         }
     }
 
-    @MainActor
     /// The map view showing the route and the user’s location.
+    @MainActor
     public var navigationMapView: NavigationMapView {
         return view as! NavigationMapView
     }
@@ -217,6 +217,10 @@ open class CarPlayMapViewController: UIViewController {
         unsubscribeFromFreeDriveNotifications()
     }
 
+    private var pointAnnotationManager: PointAnnotationManager?
+    private var pointAnnotations: [PointAnnotation] = []
+    private lazy var searchResultAnnotations = PointAnnotationsStorage<SearchResultAnnotation>()
+
     private var lifetimeSubscriptions: Set<AnyCancellable> = []
 
     func setupNavigationMapView() {
@@ -229,8 +233,22 @@ open class CarPlayMapViewController: UIViewController {
         )
 
         navigationMapView.delegate = self
-        navigationMapView.mapView.mapboxMap.onStyleLoaded.sink { [weak navigationMapView] _ in
-            navigationMapView?.localizeLabels()
+        navigationMapView.mapView.mapboxMap.onStyleLoaded.sink { [weak self] _ in
+            guard let self else { return }
+            navigationMapView.localizeLabels()
+
+            if pointAnnotationManager == nil {
+                // Creating only once becase annotations persist across style changes
+                pointAnnotationManager = navigationMapView.mapView.annotations.makePointAnnotationManager()
+
+                if pointAnnotations.count != 0,
+                   let pointAnnotationManager
+                {
+                    pointAnnotationManager.annotations = pointAnnotations
+                    pointAnnotations = []
+                }
+            }
+
         }.store(in: &lifetimeSubscriptions)
 
         navigationMapView.puckType = .puck2D(.navigationCarPlayDefault)
@@ -403,6 +421,135 @@ open class CarPlayMapViewController: UIViewController {
         }
 
         super.updateViewConstraints()
+    }
+
+    func showSearchResultsAnnotations(
+        with records: [SearchResultRecord]?,
+        selectedResult: SearchResultRecord?
+    ) {
+        guard let records else {
+            removeSearchResultsAnnotations()
+            return
+        }
+
+        let selectedImage = UIImage(
+            named: "carplay_pin_selected", in: .mapboxNavigation, with: nil
+        ) ?? UIImage()
+        let image = UIImage(
+            named: "car_play_pin_regular",
+            in: .mapboxNavigation,
+            with: nil
+        ) ?? UIImage()
+
+        do {
+            if navigationMapView.mapView.mapboxMap.image(withId: Constants.searchAnnotationImage) == nil {
+                try navigationMapView.mapView.mapboxMap.addImage(
+                    image,
+                    id: Constants.searchAnnotationImage,
+                    sdf: false,
+                    stretchX: [],
+                    stretchY: []
+                )
+            }
+            if navigationMapView.mapView.mapboxMap.image(withId: Constants.selectedSearchAnnotationImage) == nil {
+                try navigationMapView.mapView.mapboxMap.addImage(
+                    selectedImage,
+                    id: Constants.selectedSearchAnnotationImage,
+                    sdf: false,
+                    stretchX: [],
+                    stretchY: []
+                )
+            }
+        } catch {
+            Log.error(
+                "Failed to add search annotations with error: \(error.localizedDescription).",
+                category: .navigationUI
+            )
+            return
+        }
+
+        let annotationIds = searchResultAnnotations.ids
+        let currentAnnotations = pointAnnotationManager?.annotations ?? pointAnnotations
+        let remainingAnnotations = currentAnnotations.filter {
+            !annotationIds.contains($0.id)
+        }
+        searchResultAnnotations.removeAll()
+        let searchAnnotations = records.map { record -> PointAnnotation in
+
+            if record.id == selectedResult?.id {
+                var annotation = searchResultAnnotations.create(.init(searchResultRecord: record))
+                annotation.textColor = .init(.white)
+                annotation.textAnchor = .center
+                annotation.textSize = 14.0
+                annotation = annotation.textOffset(x: 0, y: -0.4)
+                annotation.image = .init(
+                    image: selectedImage,
+                    name: Constants.selectedSearchAnnotationImage
+                )
+                annotation.textField = record.indexStringValue
+                annotation.symbolSortKey = Double(records.count + 1)
+                return annotation
+            } else {
+                var annotation = searchResultAnnotations.create(.init(searchResultRecord: record))
+                annotation.textColor = .init(.white)
+                annotation.textAnchor = .center
+                annotation.textSize = 11.2
+                annotation = annotation.textOffset(x: 0, y: -0.3)
+                annotation.image = .init(
+                    image: image,
+                    name: Constants.searchAnnotationImage
+                )
+                let index = (record.serverIndex ?? 0) % records.count
+                annotation.symbolSortKey = Double(records.count - index)
+                annotation.textField = record.indexStringValue
+                return annotation
+            }
+        }
+        let annotations = searchAnnotations + remainingAnnotations
+        if let pointAnnotationManager {
+            pointAnnotationManager.annotations = annotations
+        } else {
+            pointAnnotations = annotations
+        }
+
+        fitCameraToSearchResults(searchResults: records)
+    }
+
+    func removeSearchResultsAnnotations() {
+        guard !searchResultAnnotations.isEmpty else { return }
+
+        let annotationIds = searchResultAnnotations.ids
+        pointAnnotationManager?.annotations = []
+        searchResultAnnotations.removeAll()
+        navigationMapView.navigationCamera.update(cameraState: .following)
+    }
+
+    func fitCameraToSearchResults(searchResults: [SearchResultRecord]) {
+        navigationMapView.navigationCamera.stop()
+        let initialCameraOptions = CameraOptions(
+            padding: navigationMapView.navigationCamera.viewportPadding,
+            bearing: 0,
+            pitch: 0
+        )
+        let coordinates = searchResults.map { $0.coordinate }
+        do {
+            let cameraOptions = try navigationMapView.mapView.mapboxMap.camera(
+                for: coordinates,
+                camera: initialCameraOptions,
+                coordinatesPadding: nil,
+                maxZoom: nil,
+                offset: nil
+            )
+            navigationMapView.mapView.camera.ease(to: cameraOptions, duration: 0.0)
+        } catch {
+            Log.error("Failed to fit the camera: \(error.localizedDescription)", category: .navigationUI)
+        }
+    }
+
+    private enum Constants {
+        static let identifier = "com.mapbox.navigation.carplay"
+        static let searchAnnotationImage = "\(identifier).search_annotation"
+        static let selectedSearchAnnotationImage = "\(identifier).search_annotation_selected"
     }
 }