+++ Sources/MapboxNavigationUIKit/CarPlay/CarPlayMapViewControllerDelegate.swift
@@ -1,5 +1,6 @@
 import MapboxMaps
 import MapboxNavigationCore
+import UIKit
 
 /// The ``CarPlayMapViewControllerDelegate`` protocol provides methods for reacting to events during free-drive
 /// navigation or route previewing in ``CarPlayMapViewController``.